requirements.txt
supabase==2.24.0
python-dotenv==1.2.1
openai>=2.7.2
streamlit>=1.51.0